rename: 아이템 컴포넌트를 여러페이지에서 사용하게되면서 공통 컴포넌트로 분리함
